refactor(tls): enhance certificate management and security configurations
- Implement parsed certificate storage with thread-safe access
- Add proper TLS configuration cloning and certificate validation
- Introduce stop channel initialization and safe closure mechanisms
- Refactor file monitoring with improved channel handling and cleanup
- Enhance certificate rotation with atomic operations and error handling
- Add secure password environment variable injection for services
- Update dependency versions including Go 1.26 and related libraries
- Optimize event publishing with context-based timeout management
- Improve build script with module skipping functionality

+++ integration_test.go
@@ -480,10 +480,10 @@ func (suite *RedisIntegrationTestSuite) TestConnectionPoolMetrics() {
 	assert.NotNil(suite.T(), stats)
 
 	// Validate metrics
-	assert.GreaterOrEqual(suite.T(), stats.TotalConns, uint32(0))
-	assert.GreaterOrEqual(suite.T(), stats.IdleConns, uint32(0))
-	assert.GreaterOrEqual(suite.T(), stats.Hits, uint64(0))
-	assert.GreaterOrEqual(suite.T(), stats.Timeouts, uint64(0))
+	assert.GreaterOrEqual(suite.T(), uint32(stats.TotalConns), uint32(0))
+	assert.GreaterOrEqual(suite.T(), uint32(stats.IdleConns), uint32(0))
+	assert.GreaterOrEqual(suite.T(), uint64(stats.Hits), uint64(0))
+	assert.GreaterOrEqual(suite.T(), uint64(stats.Timeouts), uint64(0))
 }
 
 // TestHealthCheck tests health check
